+++ src/internal/packager/validate/validate.go
@@ -14,6 +14,7 @@ import (
 	"github.com/defenseunicorns/zarf/src/config"
 	"github.com/defenseunicorns/zarf/src/config/lang"
 	"github.com/defenseunicorns/zarf/src/pkg/utils"
+	"github.com/defenseunicorns/zarf/src/pkg/utils/helpers"
 	"github.com/defenseunicorns/zarf/src/types"
 )
 
@@ -92,7 +93,7 @@ func ImportPackage(composedComponent *types.ZarfComponent) error {
 		if path != "" {
 			return fmt.Errorf(lang.PkgValidateErrImportOptions, composedComponent.Name)
 		}
-		ok := utils.IsOCIURL(url)
+		ok := helpers.IsOCIURL(url)
 		if !ok {
 			return fmt.Errorf(lang.PkgValidateErrImportURLInvalid, composedComponent.Import.URL)
 		}
