+++ src/pkg/utils/io.go
@@ -35,7 +35,7 @@ func GetCryptoHashFromFile(path string, hashName crypto.Hash) (string, error) {
 	var data io.ReadCloser
 	var err error
 
-	if IsURL(path) {
+	if helpers.IsURL(path) {
 		// Handle download from URL
 		message.Warn("This is a remote source. If a published checksum is available you should use that rather than calculating it directly from the remote link.")
 		data = Fetch(path)
